fix: Evade leaked min and max macro bug in MSVC
There is a bug in one or more Windows headers that makes them leak
macros with too general names, which then changes the behavior of
other code, or makes it invalid. Since this bug has been around for a
long time we cannot expect it to be fixed any time soon. This
workaround deactivates the max macro for the affected rows, however it
does not fix the problem for future code.