fix: Remove value initialization to allow complex types
SyncSData is templated to contain any type of data, however with
default the default constructor explicitly initializing the value to
zero more complex types cannot be used in the template. To allow for
complex types we fall back on implicit value initialization for the
underlying type or the client code may specify an initial value in the
other constructor.